[feat] 제작품 시간 계산
모든 건물을 매 초마다 확인하는건 비효율적입니다. 왕국의 상태가 나갈 때 시간을 기록해서 들어올 때 시간을 계산하여 제작품의 상태를 보여줍니다.

+++ Assets/13.Data/CraftData.cs
@@ -27,6 +27,6 @@ public class CraftData : ScriptableObject
     public bool IsResource => isResource;
     public ResultItem CraftResult => craftResult;
     public ResultItem[] Ingredients => ingredients;
-    public float CraftTime => craftTime;
-    public float CraftCost => craftCost;
+    public int CraftTime => (int)craftTime;
+    public int CraftCost => (int)craftCost;
 }
